% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{check_api}
\alias{check_api}
\title{Check API}
\usage{
check_api()
}
\description{
Check if the API is available and how many calls can you can use.
}
\examples{
check_api()
}
\references{
\href{https://docs.github.com/en/rest/reference/rate-limit#understanding-your-rate-limit-status}{Help page}
}
