% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{get_comments}
\alias{get_comments}
\alias{get_all_comments}
\title{Retrieve comments}
\usage{
get_comments(repository, issue)

get_all_comments(repository)
}
\arguments{
\item{repository}{A character of repository "owner/repo".}

\item{issue}{The number of the issue}
}
\value{
A \code{data.frame}
}
\description{
Tidy comments of a repository
}
\section{Functions}{
\itemize{
\item \code{get_all_comments}: The same but for all comments of a repository.
}}

\examples{
get_comments("llrs/blogR", 10)
}
