% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_events}
\alias{get_events}
\title{Retrieve events of an issue}
\usage{
get_events(repository, issue)
}
\arguments{
\item{repository}{A character of repository "owner/repo".}

\item{issue}{The number of the issue}
}
\description{
Retrieve events of an issue
}
\examples{
get_events("llrs/blogR", 2)
}
\seealso{
\code{\link[=get_issue]{get_issue()}}
}
