% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issues}
\alias{get_issues}
\title{Get issues from a repository}
\usage{
get_issues(repository)
}
\arguments{
\item{repository}{A character of form "user/repo" pointing to an existing
repository.}
}
\value{
A data.frame with the information available about the issue.
}
\description{
Download all the opening issues of a repository
}
\examples{
issues <- get_issues("llrs/blogR")
}
\seealso{
\code{\link[=get_comments]{get_comments()}}
}
