% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_labels}
\alias{get_labels}
\title{Retrieve labels}
\usage{
get_labels(repository, issue = NULL)
}
\arguments{
\item{repository}{A character of repository "owner/repo".}

\item{issue}{The number of the issue}
}
\value{
A \code{data.frame} with the information available about the issue.
}
\description{
Download all the labels of a repository or a single issue by id.
}
\examples{
issues <- get_labels("llrs/blogR")
get_labels("llrs/blogR", 11)
}
\seealso{
\code{\link[=get_comments]{get_comments()}}
}
