% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{get_pullrequests}
\alias{get_pullrequests}
\title{Retrieve only pull requests.}
\usage{
get_pullrequests(repository, issue = NULL)
}
\arguments{
\item{repository}{A character of repository "owner/repo".}

\item{issue}{The number of the issue}
}
\value{
A \code{data.frame} with the information of the pr
}
\description{
Retrieve only pull requests from a repository.
}
\examples{
get_pullrequests("llrs/BioCor", 27)
get_pullrequests("llrs/BioCor")
}
\seealso{
\code{\link[=get_comments]{get_comments()}} As they are also listed there.
}
