% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{get_timelines}
\alias{get_timelines}
\title{Retrieve timeline events}
\usage{
get_timelines(repository, issue = NULL)
}
\arguments{
\item{repository}{A character of repository "owner/repo".}

\item{issue}{The number of the issue}
}
\value{
A \code{data.frame} when finalized.
}
\description{
Retrieve timeline events
}
\examples{
get_timelines("Bioconductor/Contributions", 400)
}
