% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{link_issue}
\alias{link_issue}
\alias{link_profile}
\title{Helpers to link back to GitHub}
\usage{
link_issue(repository, n)

link_profile(id)
}
\arguments{
\item{repository}{id/repo of the repository to link (e. ex: "llrs/socialGH")}

\item{n}{Issue number}

\item{id}{Id of the user to link (e. ex: "llrs")}
}
\description{
Create links to the issue or profile, it is vectorized
}
