% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{get_pr}
\alias{get_pr}
\title{Title}
\usage{
get_pr(repository)
}
\arguments{
\item{repository}{A character of form "user/repo" pointing to an existing
repository.}
}
\value{

}
\description{
Title
}
\examples{
get_pr("llrs/BioCor")
}
